+++ src/ec4py/lsv_datas.py
@@ -92,8 +92,12 @@ class LSV_Datas:
         self.datas[item_index] = new_LSV
     #############################################################################
     
-    def __sub__(self, other: LSV_Data):
-        """_summary_
+    def __len__(self):
+        return len(self.datas)
+    
+    def __add__(self, other: LSV_Data):
+        
+        """Add to the current
 
         Args:
             other (LSV_Data): LSV_Data to be added 
@@ -102,21 +106,118 @@ class LSV_Datas:
             LSV_Datas: returns a copy of the initial dataset. 
         """
 
-        if isinstance(other, LSV_Data):
-            new_LSVs = copy.deepcopy(self)
-            for new_LSV in new_LSVs:
-                new_LSV.i = new_LSV.i - other.i
-              
-        elif isinstance(other, LSV_Datas):
-            new_LSVs = copy.deepcopy(self)
-            for new_LSV in new_LSVs:
-                new_LSV.i = new_LSV.i - other.i
+        new_LSVs = copy.deepcopy(self)
+        new_LSVs.add(other)
         return new_LSVs
+    
+    def __sub__(self, other: LSV_Data):
+        """Subtract from the current.
 
+        Args:
+            other (LSV_Data): LSV_Datas, LSV_Data or number to be subtracted 
 
+        Returns:
+            LSV_Datas: returns a copy of the initial dataset. 
+        """
+        new_LSVs = copy.deepcopy(self)
+        new_LSVs.sub(other)
+        return new_LSVs
+        
+    def __mul__(self, other: float):
+        """multiply
+
+        Args:
+            other (float): LSV_Data  or number to be added 
+
+        Returns:
+            LSV_Datas: returns a copy of the initial dataset. 
+        """
+        new_LSVs = copy.deepcopy(self)
+        new_LSVs.mul(other)
+        return new_LSVs
+    
+    def __truediv__(self, other: float):
+        """divide the current by a number.
+
+        Args:
+            other (float): number 
+
+        Returns:
+            LSV_Datas: returns a copy of the initial dataset. 
+        """
+        new_LSVs = copy.deepcopy(self)
+        new_LSVs.div(other)
+        return new_LSVs
     #############################################################################
-    def __len__(self):
-        return len(self.datas)
+
+    
+    
+    def add(self,other):
+        """add
+
+        Args:
+            other (_type_, optional): _description_. Defaults to CV_Data.
+        """
+        if isinstance(other, LSV_Datas) or isinstance(other, list):
+            if len(other) == len(self.datas):
+                for i in range(0,len(self.datas)):
+                    self.datas[i].add(other[i])
+            else:
+                raise ValueError('The data sets are not of the same length.')
+        else:
+            for data in self.datas:
+                data.add(other)  
+        ##########################################        
+    def sub(self,other):
+        """sub
+
+        Args:
+            other (_type_, optional): _description_. Defaults to CV_Data.
+        """
+        if isinstance(other, LSV_Datas) or isinstance(other, list):
+            if len(other) == len(self.datas):
+                for i in range(0,len(self.datas)):
+                    self.datas[i].sub(other[i])
+            else:
+                raise ValueError('The data sets are not of the same length.')
+        else:
+            for data in self.datas:
+                data.sub(other)
+                
+    def mul(self,other):
+        """multiply the current by a number.
+
+        Args:
+            other (_type_, optional): _description_. Defaults to CV_Data.
+        """
+        if isinstance(other, LSV_Datas) or isinstance(other, list):
+            if len(other) == len(self.datas):
+                for i in range(0,len(self.datas)):
+                    self.datas[i].mul(other[i])
+            else:
+                raise ValueError('The data sets are not of the same length.')
+        else:
+            for data in self.datas:
+                data.mul(other)
+            
+    def div(self,other):
+        """divide the current by a number.
+
+        Args:
+            other (_type_, optional): _description_. Defaults to CV_Data.
+        """
+        if isinstance(other, LSV_Datas) or isinstance(other, list):
+            if len(other) == len(self.datas):
+                for i in range(0,len(self.datas)):
+                    self.datas[i].div(other[i])
+            else:
+                raise ValueError('The data sets are not of the same length.')
+        else:
+            for data in self.datas:
+                data.div(other)  
+        
+    
+    
     
     def append(self,LSV = LSV_Data):
         self.datas.append(LSV)
@@ -132,6 +233,18 @@ class LSV_Datas:
             rates.append(lsv.rate)
         return rates
     
+    
+    def set_i_at_E_to_zero(self, E:float, *args, **kwargs):
+        """Set the current at a specific voltage to zero.
+        
+        Args:
+            E (float): potential where to set the current to zero. 
+
+        """
+        for x in self.datas:
+            x.set_i_at_E_to_zero(E,*args,**kwargs)
+            
+    ################################################################
     def bg_corr(self, bg: LSV_Data|Path) -> LSV_Data:
         """Background correct the data by subtracting the bg.